+++ ziggurat/src/ieee_802154.rs
@@ -1,53 +1,54 @@
 #![allow(dead_code)]
 
+use crate::types::{NWK, EUI64};
 use std::convert::TryFrom;
 
 
 #[derive(Debug, PartialEq, Copy, Clone)]
-pub enum FrameType {
+pub enum Ieee802154FrameType {
     Beacon = 0b000,
     Data = 0b001,
     Command = 0b011,
     Ack = 0b010,
 }
 
-impl TryFrom<u8> for FrameType {
+impl TryFrom<u8> for Ieee802154FrameType {
     type Error = &'static str;
 
     fn try_from(value: u8) -> Result<Self, Self::Error> {
         match value {
-            0b000 => Ok(FrameType::Beacon),
-            0b001 => Ok(FrameType::Data),
-            0b011 => Ok(FrameType::Command),
-            0b010 => Ok(FrameType::Ack),
+            0b000 => Ok(Ieee802154FrameType::Beacon),
+            0b001 => Ok(Ieee802154FrameType::Data),
+            0b011 => Ok(Ieee802154FrameType::Command),
+            0b010 => Ok(Ieee802154FrameType::Ack),
             _ => Err("Invalid frame type"),
         }
     }
 }
 
 #[derive(Debug, PartialEq, Copy, Clone)]
-pub enum AddressingMode {
+pub enum Ieee802154AddressingMode {
     None = 0b00,
     Short = 0b10,
     Long = 0b11,
 }
 
-impl TryFrom<u8> for AddressingMode {
+impl TryFrom<u8> for Ieee802154AddressingMode {
     type Error = &'static str;
 
     fn try_from(value: u8) -> Result<Self, Self::Error> {
         match value {
-            0b00 => Ok(AddressingMode::None),
-            0b10 => Ok(AddressingMode::Short),
-            0b11 => Ok(AddressingMode::Long),
+            0b00 => Ok(Ieee802154AddressingMode::None),
+            0b10 => Ok(Ieee802154AddressingMode::Short),
+            0b11 => Ok(Ieee802154AddressingMode::Long),
             _ => Err("Invalid addressing mode"),
         }
     }
 }
 
 #[derive(Debug)]
-pub struct FrameControl {
-    pub frame_type: FrameType,
+pub struct Ieee802154FrameControl {
+    pub frame_type: Ieee802154FrameType,
     pub security_enabled: bool,
     pub frame_pending: bool,
     pub ack_request: bool,
@@ -55,20 +56,20 @@ pub struct FrameControl {
     pub reserved: bool,
     pub sequence_number_suppression: bool,
     pub information_elements_present: bool,
-    pub dest_addr_mode: AddressingMode,
+    pub dest_addr_mode: Ieee802154AddressingMode,
     pub frame_version: u8,
-    pub src_addr_mode: AddressingMode,
+    pub src_addr_mode: Ieee802154AddressingMode,
 }
 
-impl FrameControl {
+impl Ieee802154FrameControl {
     pub fn deserialize(bytes: &[u8]) -> Result<(Self, &[u8]), &'static str> {
         if bytes.len() < 2 {
-            return Err("Not enough data to parse FrameControl");
+            return Err("Not enough data to parse Ieee802154FrameControl");
         }
 
         Ok(
             (Self {
-                frame_type: FrameType::try_from(bytes[0] & 0b0000_0111)?,
+                frame_type: Ieee802154FrameType::try_from(bytes[0] & 0b0000_0111)?,
                 security_enabled: (bytes[0] & 0b0000_1000) != 0,
                 frame_pending: (bytes[0] & 0b0001_0000) != 0,
                 ack_request: (bytes[0] & 0b0010_0000) != 0,
@@ -76,9 +77,9 @@ impl FrameControl {
                 reserved: (bytes[0] & 0b1000_0000) != 0,
                 sequence_number_suppression: (bytes[1] & 0b0000_0001) != 0,
                 information_elements_present: (bytes[1] & 0b0000_0010) != 0,
-                dest_addr_mode: AddressingMode::try_from((bytes[1] >> 2) & 0b0000_0011)?,
+                dest_addr_mode: Ieee802154AddressingMode::try_from((bytes[1] >> 2) & 0b0000_0011)?,
                 frame_version: (bytes[1] >> 4) & 0b0000_0011,
-                src_addr_mode: AddressingMode::try_from((bytes[1] >> 6) & 0b0000_0011)?,
+                src_addr_mode: Ieee802154AddressingMode::try_from((bytes[1] >> 6) & 0b0000_0011)?,
             }, &bytes[2..])
         )
     }
@@ -108,7 +109,7 @@ impl FrameControl {
 }
 
 #[derive(Debug, PartialEq, Copy, Clone)]
-pub enum CommandId {
+pub enum Ieee802154CommandId {
     NotAMacCommand = 0x00,
     AssociationRequest = 0x01,
     AssociationResponse = 0x02,
@@ -121,67 +122,47 @@ pub enum CommandId {
     GtsRequest = 0x09,
 }
 
-impl TryFrom<u8> for CommandId {
+impl TryFrom<u8> for Ieee802154CommandId {
     type Error = &'static str;
 
     fn try_from(value: u8) -> Result<Self, Self::Error> {
         match value {
-            0x00 => Ok(CommandId::NotAMacCommand),
-            0x01 => Ok(CommandId::AssociationRequest),
-            0x02 => Ok(CommandId::AssociationResponse),
-            0x03 => Ok(CommandId::DisassociationNotification),
-            0x04 => Ok(CommandId::DataRequest),
-            0x05 => Ok(CommandId::PanIdConflictNotification),
-            0x06 => Ok(CommandId::OrphanNotification),
-            0x07 => Ok(CommandId::BeaconRequest),
-            0x08 => Ok(CommandId::CoordinatorRealignment),
-            0x09 => Ok(CommandId::GtsRequest),
+            0x00 => Ok(Ieee802154CommandId::NotAMacCommand),
+            0x01 => Ok(Ieee802154CommandId::AssociationRequest),
+            0x02 => Ok(Ieee802154CommandId::AssociationResponse),
+            0x03 => Ok(Ieee802154CommandId::DisassociationNotification),
+            0x04 => Ok(Ieee802154CommandId::DataRequest),
+            0x05 => Ok(Ieee802154CommandId::PanIdConflictNotification),
+            0x06 => Ok(Ieee802154CommandId::OrphanNotification),
+            0x07 => Ok(Ieee802154CommandId::BeaconRequest),
+            0x08 => Ok(Ieee802154CommandId::CoordinatorRealignment),
+            0x09 => Ok(Ieee802154CommandId::GtsRequest),
             _ => Err("Invalid addressing mode"),
         }
     }
 }
 
 #[derive(Debug, PartialEq, Copy, Clone)]
-pub struct NWK(pub u16);
-
-impl NWK {
-    pub fn to_bytes(&self) -> [u8; 2] {
-        self.0.to_le_bytes()
-    }
-}
-
-
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub struct EUI64(pub [u8; 8]);
-
-impl EUI64 {
-    pub fn to_bytes(&self) -> [u8; 8] {
-        self.0
-    }
-}
-
-
-#[derive(Debug, PartialEq, Copy, Clone)]
-pub enum Address {
+pub enum Ieee802154Address {
     NWK(NWK),
     EUI64(EUI64),
 }
 
 
 #[derive(Debug)]
-pub struct Frame {
-    pub frame_control: FrameControl,
+pub struct Ieee802154Frame {
+    pub frame_control: Ieee802154FrameControl,
     pub sequence_number: Option<u8>,
     pub dest_pan_id: Option<u16>,
-    pub dest_address: Option<Address>,
+    pub dest_address: Option<Ieee802154Address>,
     pub src_pan_id: Option<u16>,
-    pub src_address: Option<Address>,
+    pub src_address: Option<Ieee802154Address>,
     pub payload: Vec<u8>,
     pub fcs: u16,
 }
 
 
-impl Frame {
+impl Ieee802154Frame {
     pub fn from_bytes(data: &[u8]) -> Result<Self, &'static str> {
         if data.len() < 2 + 2 {
             return Err("Data too short to contain a frame");
@@ -197,7 +178,7 @@ impl Frame {
         let mut offset = 0;
 
         // Parse frame control
-        let (frame_control, remaining) = FrameControl::deserialize(data)?;
+        let (frame_control, remaining) = Ieee802154FrameControl::deserialize(data)?;
 
         // Parse sequence number
         let sequence_number = if frame_control.sequence_number_suppression {
@@ -210,11 +191,11 @@ impl Frame {
 
         // Parse destination PAN ID and address
         let (dest_pan_id, dest_address) = match frame_control.dest_addr_mode {
-            AddressingMode::Short => {
+            Ieee802154AddressingMode::Short => {
                 let pan_id = u16::from_le_bytes([remaining[offset], remaining[offset + 1]]);
                 offset += 2;
 
-                let address = Address::NWK(NWK(u16::from_le_bytes([
+                let address = Ieee802154Address::NWK(NWK(u16::from_le_bytes([
                     remaining[offset],
                     remaining[offset + 1],
                 ])));
@@ -222,11 +203,11 @@ impl Frame {
 
                 (Some(pan_id), Some(address))
             }
-            AddressingMode::Long => {
+            Ieee802154AddressingMode::Long => {
                 let pan_id = u16::from_le_bytes([remaining[offset], remaining[offset + 1]]);
                 offset += 2;
 
-                let address = Address::EUI64(EUI64([
+                let address = Ieee802154Address::EUI64(EUI64([
                     remaining[offset],
                     remaining[offset + 1],
                     remaining[offset + 2],
@@ -240,13 +221,13 @@ impl Frame {
 
                 (Some(pan_id), Some(address))
             }
-            AddressingMode::None => (None, None),
+            Ieee802154AddressingMode::None => (None, None),
         };
 
         // Parse source PAN ID
         let src_pan_id = if frame_control.pan_id_compression {
             dest_pan_id
-        } else if (frame_control.frame_type == FrameType::Data) || (frame_control.frame_type == FrameType::Command) {
+        } else if (frame_control.frame_type == Ieee802154FrameType::Data) || (frame_control.frame_type == Ieee802154FrameType::Command) {
             let pan_id = u16::from_le_bytes([remaining[offset], remaining[offset + 1]]);
             offset += 2;
             Some(pan_id)
@@ -256,16 +237,16 @@ impl Frame {
 
         // Parse source address
         let src_address = match frame_control.src_addr_mode {
-            AddressingMode::Short => {
-                let address = Address::NWK(NWK(u16::from_le_bytes([
+            Ieee802154AddressingMode::Short => {
+                let address = Ieee802154Address::NWK(NWK(u16::from_le_bytes([
                     remaining[offset],
                     remaining[offset + 1],
                 ])));
                 offset += 2;
                 Some(address)
             }
-            AddressingMode::Long => {
-                let address = Address::EUI64(EUI64([
+            Ieee802154AddressingMode::Long => {
+                let address = Ieee802154Address::EUI64(EUI64([
                     remaining[offset],
                     remaining[offset + 1],
                     remaining[offset + 2],
@@ -313,8 +294,8 @@ impl Frame {
         }
         if let Some(address) = &self.dest_address {
             data.extend(match address {
-                Address::NWK(addr) => addr.to_bytes().to_vec(),
-                Address::EUI64(addr) => addr.to_bytes().to_vec(),
+                Ieee802154Address::NWK(addr) => addr.to_bytes().to_vec(),
+                Ieee802154Address::EUI64(addr) => addr.to_bytes().to_vec(),
             });
         }
 
@@ -327,8 +308,8 @@ impl Frame {
 
         if let Some(address) = &self.src_address {
             data.extend(match address {
-                Address::NWK(addr) => addr.to_bytes().to_vec(),
-                Address::EUI64(addr) => addr.to_bytes().to_vec(),
+                Ieee802154Address::NWK(addr) => addr.to_bytes().to_vec(),
+                Ieee802154Address::EUI64(addr) => addr.to_bytes().to_vec(),
             });
         }
