+++ ziggurat/tests/frame_parsing.rs
@@ -1,5 +1,8 @@
-use zzigbee_frames::*;
 use hex_literal::hex;
+
+use ziggurat::types::*;
+use ziggurat::zigbee_nwk::*;
+use ziggurat::zigbee_aps::*;
 
 
 #[test]
