--- ziggurat/z802154/Cargo.toml
@@ -1,5 +1,0 @@
-[package]
-name = "z802154"
-version = "0.0.1"
-authors = []
-edition = "2021"
