--- ziggurat/zzigbee_frames/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "zzigbee_frames"
-version = "0.0.1"
-authors = []
-edition = "2021"
-
-[dependencies]
-"aes" = "=0.9.0-pre.2"
-"cbc" = "=0.2.0-pre.2"
-"constant_time_eq" = "=0.3.1"
-"hex-literal" = "=0.4.1"
-"hex" = "=0.4.3"
